+++ sdk/src/main/java/com/hedera/hashgraph/sdk/ContractCreateFlow.java
@@ -95,6 +95,8 @@ public class ContractCreateFlow implements WithExecute<TransactionResponse> {
     private int maxAutomaticTokenAssociations = 0;
     @Nullable
     private Duration autoRenewPeriod = null;
+    @Nullable
+    private AccountId autoRenewAccountId = null;
     private byte[] constructorParameters = {};
     @Nullable
     private String contractMemo = null;
@@ -309,6 +311,17 @@ public class ContractCreateFlow implements WithExecute<TransactionResponse> {
         return this;
     }
 
+    @Nullable
+    public AccountId getAutoRenewAccountId() {
+        return autoRenewAccountId;
+    }
+
+    public ContractCreateFlow setAutoRenewAccountId(AccountId autoRenewAccountId) {
+        Objects.requireNonNull(autoRenewAccountId);
+        this.autoRenewAccountId = autoRenewAccountId;
+        return this;
+    }
+
     /**
      * Extract the byte string representation.
      *
@@ -504,6 +517,9 @@ public class ContractCreateFlow implements WithExecute<TransactionResponse> {
         }
         if (autoRenewPeriod != null) {
             contractCreateTx.setAutoRenewPeriod(autoRenewPeriod);
+        }
+        if (autoRenewAccountId != null) {
+            contractCreateTx.setAutoRenewAccountId(autoRenewAccountId);
         }
         if (contractMemo != null) {
             contractCreateTx.setContractMemo(contractMemo);